Implement q_remove_head, q_remove_tail
These functions implement the removal of elements from the head and
tail of the queue respectively. They first check if the queue is empty
using list_empty. Then, they use list_entry to retrieve the list
element from the head or tail. The element value is copied to sp using
strncpy, and the node is removed from the list using list_del_init.

Return the element object if the list is not empty; otherwise,
return NULL.

In the future, consider refactoring q_remove_head and q_remove_tail as
they share similar code.