Delete middle node in queue
This commit implements a function to delete the middle node in a
queue. Before deletion, the function checks the queue if the head is
NULL or if the list is empty. The function iterates through the queue
using two pointers,one moving twice as fast as the other. When the
fast pointer reaches the end, the slow pointer points to the middle
node. The middle node is then removed from the queue.

A potential future improvement regarding the utilization of an
indirect pointer.